Revert "Revert "lex MinInt; store Integer as Long in token""
This reverts commit 0ad9b2b8ec6dac64751e9f54a2cc2ae3165ad63d.

+++ src/main/java/mtm68/ast/nodes/IntLiteral.java
@@ -1,12 +1,10 @@
 package mtm68.ast.nodes;
 
-import java.math.BigInteger;
-
 import edu.cornell.cs.cs4120.util.SExpPrinter;
 
-public class IntLiteral extends Literal<BigInteger> {
+public class IntLiteral extends Literal<Long> {
 	
-	public IntLiteral(BigInteger value) {
+	public IntLiteral(Long value) {
 		super(value);
 	}
